Load linux zsh aliases conditionally

+++ zsh/aliases.zsh
@@ -3,7 +3,7 @@
 alias rl='source ~/.zshrc; echo ".zshrc reloaded"'
 alias regen='zgen reset;source ~/.zshrc'
 
-alias ls='ls -p'
+alias ls='ls -GFHp'
 alias ll='ls -la'
 
 # Easier navigation
@@ -22,7 +22,7 @@ alias -g H='| head'
 alias -g T='| tail'
 alias -g G='| grep'
 alias -g NUL='&> /dev/null'
-alias -g C='| xclip -selection c'
+alias -g C='| pbcopy'
 alias -g J='| python -m json.tool'
 
 # Prompt if overwriting
@@ -30,16 +30,25 @@ alias cp='cp -i'
 alias mv='mv -i'
 
 alias sudo='sudo '
-alias sagi='sudo apt-get install'
-alias sagu='sudo apt-get update'
-alias saar='sudo add-apt-repository'
-alias sagr='sudo apt-get remove'
+
 hs(){ history | grep -i "$1" ;}
 alias .f='cd ~/.dotfiles'
 alias fd='find . -type d -name'
 alias ff='find . -type f -name'
-alias pbcopy='xclip -selection c'
-alias pbpaste='xclip -selection clipboard -o'
+
+
+
+if [ "$(uname -s)" = "Linux" ];then
+	alias ls='ls --color=auto -p'
+	alias sagi='sudo apt-get install'
+	alias sagu='sudo apt-get update'
+	alias saar='sudo add-apt-repository'
+	alias sagr='sudo apt-get remove'
+	alias -g C='| xclip -selection c'
+  alias pbcopy='xclip -selection c'
+	alias pbpaste='xclip -selection clipboard -o'
+	alias rsx='sudo killall xbindkeys && sudo xbindkeys && echo "xbindkeys restarted"'
+fi
 
 # Other bash stuff
 alias mkcd='mkc'
@@ -54,7 +63,7 @@ alias be="bundle exec"
 alias mm="middleman"
 alias djs="dashing-js start"
 alias subla="subl . -a"
-alias rsx='sudo killall xbindkeys && sudo xbindkeys && echo "xbindkeys restarted"'
+
 
 # Zshmarks ==> Bashmarks
 alias g="jump"
